moved is_binary, is_structured into http module
